+++ app/src/androidTest/java/woowacourse/movie/booking/BookingActivityTest.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie
+package woowacourse.movie.booking
 
 import android.content.Intent
 import androidx.test.core.app.ActivityScenario
@@ -20,7 +20,8 @@ import org.junit.Before
 import org.junit.Test
 import org.junit.jupiter.api.DisplayName
 import org.junit.runner.RunWith
-import woowacourse.movie.model.Movie
+import woowacourse.movie.R
+import woowacourse.movie.model.data.Movie
 import woowacourse.movie.ui.view.booking.BookingActivity
 import woowacourse.movie.ui.view.booking.BookingSummaryActivity
 import java.time.LocalDate
