+++ app/src/main/java/woowacourse/movie/model/data/Movie.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model
+package woowacourse.movie.model.data
 
 import java.io.Serializable
 import java.time.LocalDate
