+++ app/src/main/java/woowacourse/movie/model/policy/DefaultPricingPolicy.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model
+package woowacourse.movie.model.policy
 
 class DefaultPricingPolicy : PricingPolicy {
     override fun calculatePrice(headCount: Int): Int = headCount * DEFAULT_PRICE
