+++ app/src/main/java/woowacourse/movie/model/schedule/DayType.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.model
+package woowacourse.movie.model.schedule
 
 import java.time.DayOfWeek
 import java.time.LocalDate
